# public/swagger.json
{
  "swagger": "2.0",
  "info": {
     "version": "1.0",
     "title": "La Villa Pasta REST API",
     "description": "Backend REST API for La Villa Pasta to manage items, orders and users",
     "contact": {}
  },
  "host": "la-villa-pasta-backend.herokuapp.com",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": [
     "https"
  ],
  "consumes": [
     "application/json"
  ],
  "produces": [
     "application/json"
  ],
  "paths": {
...
     "/billings/604fee5304af8392212f40d8": {
        "put": {
           "description": "Updates User Billing in database",
           "summary": "Update Billing",
           "tags": [
              "Billing"
           ],
           "operationId": "UpdateBilling",
           "deprecated": false,
           "produces": [
              "application/json"
           ],
           "consumes": [
              "application/json"
           ],
           "parameters": [
              {
                 "name": "Authorization",
                 "in": "header",
                 "required": false,
                 "default": "Bearer {token}",
                 "type": "string"
              },
              {
                 "name": "Content-Type",
                 "in": "header",
                 "required": true,
                 "type": "string",
                 "description": "JSON Type"
              },
              {
                 "name": "Body",
                 "in": "body",
                 "required": true,
                 "description": "",
                 "schema": {
                    "$ref": "#/definitions/UpdateBillingRequest"
                 }
              }
           ],
           "responses": {
              "200": {
                 "description": "",
                 "headers": {}
              }
           }
        }
     },
     "/billings/604408c88e85f6720bd7d221": {
        "delete": {
           "description": "Deletes User Billing from database",
           "summary": "Delete Billing",
           "tags": [
              "Billing"
           ],
           "operationId": "DeleteBilling",
           "deprecated": false,
           "produces": [
              "application/json"
           ],
           "parameters": [
              {
                 "name": "Authorization",
                 "in": "header",
                 "required": false,
                 "default": "Bearer {token}",
                 "type": "string"
              }
           ],
           "responses": {
              "200": {
                 "description": "",
                 "headers": {}
              }
           }
        }
     },
...
  },
  "definitions": {
...
     "UpdateBillingRequest": {
        "title": "UpdateBillingRequest",
        "example": {
           "address": "240 Bloor St East, Toronto, ON"
        },
        "type": "object",
        "properties": {
           "address": {
              "type": "string"
           }
        },
        "required": [
           "address"
        ]
     },
...
  },
...
}